move linking options to shared module
