+++ toolchain/driver/BUILD
@@ -150,6 +150,10 @@ cc_library(
         "format_subcommand.h",
         "language_server_subcommand.cpp",
         "language_server_subcommand.h",
+        "link_driver.cpp",
+        "link_driver.h",
+        "link_options.cpp",
+        "link_options.h",
         "link_subcommand.cpp",
         "link_subcommand.h",
         "lld_subcommand.cpp",
